fix: fix temp issues

+++ src/domain/git/coco_commit.rs
@@ -1,3 +1,5 @@
+use serde::{Deserialize, Serialize};
+
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct CocoCommit {
     pub rev: String,
@@ -9,8 +11,8 @@ pub struct CocoCommit {
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct FileChange {
-    pub added: int,
-    pub deleted: int,
+    pub added: i64,
+    pub deleted: i64,
     pub file: String,
     pub mode: String
 }
